model for notification created
