+++ forum/forum/settings.py
@@ -119,6 +119,7 @@ INSTALLED_APPS = [
     'dashboard',
     'storages',
     'drf_yasg',
+    'notifications',
 ]
 
 MIDDLEWARE = [
@@ -279,6 +280,5 @@ AWS_STORAGE_BUCKET_NAME = os.environ.get('AWS_STORAGE_BUCKET_NAME')
 AWS_S3_REGION_NAME = os.environ.get('AWS_S3_REGION_NAME')
 
 # S3 Storage for static and media files
-STATICFILES_STORAGE =  'forum.storages.StaticStorage'
+STATICFILES_STORAGE = 'forum.storages.StaticStorage'
 DEFAULT_FILE_STORAGE = 'forum.storages.MediaStorage'
-
